Change HTML title when adding a bookmark
I know, it's not a great quickfix but I did not want to bother
understanding the templating system. So copying the file works for now.

Why: When adding a bookmark via the bookmarklet, it opens a new popup
window. As I'm using a tiling window manager, I want to do reposition
the window automatically and, therefore, I need to have attributes to
match the window. The window title and app_id usually work well. With
this change the window title when adding a bookmark is now different, so
I can apply changes to the window (and only when adding, not browsing my
bookmarks).